:lock: ToDoList 로그인한 유저가 다른 유저의 정보를 요청하는 경우 ResponseStatusException

+++ ZZANDI/src/main/java/com/ll/zzandi/controller/ToDoListController.java
@@ -6,11 +6,12 @@ import com.ll.zzandi.dto.ToDoListDto;
 import com.ll.zzandi.enumtype.ToDoType;
 import com.ll.zzandi.service.ToDoListService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import java.util.List;
+import org.springframework.web.server.ResponseStatusException;
 
 @Controller
 @RequiredArgsConstructor
@@ -23,6 +24,7 @@ public class ToDoListController {
         return "todo/ToDoListMainAsync";
     }
 
+
     @GetMapping("/create")
     @ResponseBody
     public ToDoList createToDo(@AuthenticationPrincipal User user, String content) {
@@ -33,35 +35,55 @@ public class ToDoListController {
 
     @GetMapping("/update")
     @ResponseBody
-    public ToDoList updateToDo(@RequestParam long id, @RequestParam String content) {
+    public ToDoList updateToDo(@RequestParam long id, @RequestParam String content, @AuthenticationPrincipal User user) {
+        ToDoList toDoById = toDoListService.findById(id);
 
+        if (! toDoById.getUser().getId().equals(user.getId())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+        }
 
         return toDoListService.update(id, content);
     }
 
     @GetMapping("/change")
     @ResponseBody
-    public ToDoList changeToDoType (long id) {
+    public ToDoList changeToDoType (@RequestParam long id, @AuthenticationPrincipal User user) {
+        ToDoList toDoById = toDoListService.findById(id);
+
+        if (! toDoById.getUser().getId().equals(user.getId())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+        }
+
         return toDoListService.changeType(id);
     }
 
     @DeleteMapping("/delete")
     @ResponseBody
-    public void deleteToDo (long id) {
+    public void deleteToDo (@RequestParam long id, @AuthenticationPrincipal User user) {
+        ToDoList toDoById = toDoListService.findById(id);
+
+        if (! toDoById.getUser().getId().equals(user.getId())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+        }
+
         toDoListService.delete(id);
     }
 
     @GetMapping("/list-data")
     @ResponseBody
     public List<ToDoList> ToDoToJson(@RequestParam(required = false) ToDoType type, @AuthenticationPrincipal User user) {
-//        return (type == null) ? toDoListService.findAll() : toDoListService.findAllByType(type);
-
         return (type == null) ? toDoListService.findAllByUser(user) : toDoListService.findAllByUserAndType(user, type);
     }
 
     @GetMapping("/todo-data")
     @ResponseBody
-    public ToDoList ToDoToJson(@RequestParam(required = false) long id) {
-        return toDoListService.findById(id);
+    public ToDoList ToDoToJson(@RequestParam(required = false) long id, @AuthenticationPrincipal User user) {
+        ToDoList toDoById = toDoListService.findById(id);
+
+        if (! toDoById.getUser().getId().equals(user.getId())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+        }
+
+        return toDoById;
     }
 }
